pyproject.toml
[tool.poetry]
name = "flask-keycloak-oidc"
version = "0.1.0"
description = ""
authors = ["David Knoebl"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
flask-oidc = "^2.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]